@@ -3,22 +3,24 @@
   "private": true,
   "dependencies": {
     "@fontsource/roboto": "^4.5.8",
+    "@lezer/highlight": "^1.1.3",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
     "@material-ui/styles": "^4.11.5",
-    "@uiw/react-codemirror": "^3.1.0",
+    "@uiw/codemirror-themes": "^4.19.5",
+    "@uiw/react-codemirror": "^4.19.5",
     "axios": "^0.27.2",
-    "codemirror": "^5.62.3",
+    "codemirror": "^6.0.1",
     "date-fns": "^2.29.2",
     "history": "^5.3.0",
     "ipaddr.js": "^2.0.1",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-data-table-component": "^6.11.8",
-    "react-dom": "^17.0.2",
-    "react-is": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react": "^18.2.0",
+    "react-data-table-component": "^7.5.3",
+    "react-dom": "^18.2.0",
+    "react-is": "^18.2.0",
+    "react-router-dom": "^6.6.1",
+    "react-scripts": "5.0.1",
     "react-use": "^17.4.0",
     "styled-components": "^5.3.5"
   },
